+++ src/contentful/types.ts
@@ -1,3 +1,5 @@
+import { Entry } from "contentful";
+
 export interface PageTreeNode {
     id: string;
     slug: string | null;
@@ -24,22 +26,6 @@ export interface BaseContentfulEntry {
     };
 }
 
-export interface DetailedContentfulEntry<T> {
-    sys: {
-        id: string;
-        type: string;
-        createdAt: string;
-        updatedAt: string;
-        locale: string;
-        contentType: BaseContentfulEntry;
-    };
-    fields: T;
-}
-
-export interface ContentfulEntry<T> extends BaseContentfulEntry {
-    fields?: T;
-}
-
 export interface ContentfulBasePage {
     slug: string;
     childPages: BaseContentfulEntry[];
@@ -47,8 +33,8 @@ export interface ContentfulBasePage {
 
 export interface ContentfulPage extends ContentfulBasePage {
     title: string;
-    seo: DetailedContentfulEntry<ContentfulPageSeo>;
-    content: DetailedContentfulEntry<Record<string, any>>;
+    seo: Entry<ContentfulPageSeo>;
+    content: Entry<Record<string, any>>;
 }
 
 export interface ContentfulPageSeo {
@@ -63,42 +49,5 @@ export interface PageProps {
         destination: string;
         permanent: boolean;
     };
-    page?: {
-        id: string;
-        path: string;
-        publishState: "published" | string;
-        timeCreated: string;
-        slug: string | null;
-        metaTitle: string | null;
-        metaDescription: string | null;
-        locale: string;
-        name: string;
-        headline: string | null;
-        type: string;
-        properties: PageProperties;
-        content: PageContent;
-    };
-}
-
-export interface PageProperties {
-    [key: string]: any;
-}
-
-export interface PageContent {
-    slots: Record<string, SlotData>;
-}
-
-export interface SlotData {
-    blocks: BlockData[];
-}
-
-export interface BlockData {
-    id: string;
-    anchor: string;
-    anchorLabel: string;
-    type: string;
-    timeCreated: string;
-    timeModified: string;
-    content: any;
-    hidden: boolean;
+    page?: Entry<ContentfulPage>;
 }